UserProject added. Much of the algorithms in this user project will be added to the framework.
In the first architecture design of the framework all the algorithms
were mplemented by the user, as in this UserProject. But on a later
version of the URSA framework these algorithms will be built-in the
framework and the user still will have the possibility to write his own
algorithms for clustering and analysis.