Some minor changes in Document.java, DataCluster.java and DataObject.java. ome new get/set functions were added, as well as a new variable index in DataObject.java
DocumentClustering.java was adapted to support the new changes in
DataObject.java. The new variable "index" is needed for the constructor
of a new DataObject object.

BestStarClusteringtrategyOld is the old algorithm. It work, but was hard
to understand. Will be active until all the tests with the new version
are completed.